Position is now updated via input component

+++ common/src/main/kotlin/org/sandboxpowered/silica/component/PositionComponent.kt
@@ -1,12 +1,13 @@
 package org.sandboxpowered.silica.component
 
 import com.artemis.PooledComponent
+import org.joml.Vector3d
 import org.joml.Vector3f
 
 class PositionComponent : PooledComponent() {
-    val pos: Vector3f = Vector3f()
+    val pos: Vector3d = Vector3d()
 
     override fun reset() {
-        pos.set(0f, 0f, 0f)
+        pos.set(0.0, 0.0, 0.0)
     }
 }
